+++ src/main/java/com/eshop/entity/collect/ClProductAttrVal.java
@@ -14,7 +14,7 @@ import lombok.ToString;
  * </p>
  *
  * @author collect
- * @since 2025-02-11
+ * @since 2025-02-12
  */
 @Getter
 @Setter
@@ -25,12 +25,12 @@ public class ClProductAttrVal implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 产品id
      */
-    private Integer productId;
+    private Long productId;
 
     /**
      * 产品编码
@@ -40,7 +40,7 @@ public class ClProductAttrVal implements Serializable {
     /**
      * 属性Id
      */
-    private Integer attrId;
+    private Long attrId;
 
     /**
      * 属性值名称
@@ -50,7 +50,7 @@ public class ClProductAttrVal implements Serializable {
     /**
      * 属性值id
      */
-    private Integer attrValId;
+    private Long attrValId;
 
     /**
      * 属性值名称
@@ -65,7 +65,7 @@ public class ClProductAttrVal implements Serializable {
     /**
      * 属性值图片_id
      */
-    private Integer imgId;
+    private Long imgId;
 
     /**
      * 自定义属性值
