+++ src/main/java/com/eshop/entity/config/TbCountryExpand.java
@@ -14,7 +14,7 @@ import lombok.ToString;
  * </p>
  *
  * @author config
- * @since 2025-02-11
+ * @since 2025-02-12
  */
 @Getter
 @Setter
@@ -25,9 +25,9 @@ public class TbCountryExpand implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
-    private Integer countryId;
+    private Long countryId;
 
     /**
      * 国家英文名
