+++ src/main/java/com/eshop/entity/config/TbRoleDetail.java
@@ -14,7 +14,7 @@ import lombok.ToString;
  * </p>
  *
  * @author config
- * @since 2025-02-11
+ * @since 2025-02-12
  */
 @Getter
 @Setter
@@ -25,15 +25,15 @@ public class TbRoleDetail implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 角色编号
      */
-    private Integer roleid;
+    private Long roleid;
 
     /**
      * 菜单编号
      */
-    private Integer menuid;
+    private Long menuid;
 }
