+++ src/main/java/com/eshop/entity/config/TbShopDepot.java
@@ -14,7 +14,7 @@ import lombok.ToString;
  * </p>
  *
  * @author config
- * @since 2025-02-11
+ * @since 2025-02-12
  */
 @Getter
 @Setter
@@ -25,15 +25,15 @@ public class TbShopDepot implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 店铺_id
      */
-    private Integer shopId;
+    private Long shopId;
 
     /**
      * 仓库_id
      */
-    private Integer depotId;
+    private Long depotId;
 }
