+++ src/main/java/com/eshop/entity/depot/TbFlexibleCustomizeMaterialDetail.java
@@ -14,7 +14,7 @@ import lombok.ToString;
  * </p>
  *
  * @author depot
- * @since 2025-02-11
+ * @since 2025-02-12
  */
 @Getter
 @Setter
@@ -25,12 +25,12 @@ public class TbFlexibleCustomizeMaterialDetail implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 柔性定制item列表id
      */
-    private Integer fcsId;
+    private Long fcsId;
 
     /**
      * 批次号
