+++ src/main/java/com/eshop/entity/depot/TbStock.java
@@ -15,7 +15,7 @@ import lombok.ToString;
  * </p>
  *
  * @author depot
- * @since 2025-02-11
+ * @since 2025-02-12
  */
 @Getter
 @Setter
@@ -26,12 +26,12 @@ public class TbStock implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 仓库id
      */
-    private Integer depotId;
+    private Long depotId;
 
     /**
      * 仓位编码
@@ -51,27 +51,27 @@ public class TbStock implements Serializable {
     /**
      * 库存数量
      */
-    private Integer inventoryQuantity;
+    private Long inventoryQuantity;
 
     /**
      * 缺货数量
      */
-    private Integer stockoutQuantity;
+    private Long stockoutQuantity;
 
     /**
      * 采购中数量
      */
-    private Integer purchaseQuantity;
+    private Long purchaseQuantity;
 
     /**
      * 待处理数量
      */
-    private Integer pendingQuantity;
+    private Long pendingQuantity;
 
     /**
      * 冻结数量
      */
-    private Integer freezeQuantity;
+    private Long freezeQuantity;
 
     /**
      * 最低采购单价
