+++ src/main/java/com/eshop/entity/depot/TbStorageDetail.java
@@ -15,7 +15,7 @@ import lombok.ToString;
  * </p>
  *
  * @author depot
- * @since 2025-02-11
+ * @since 2025-02-12
  */
 @Getter
 @Setter
@@ -26,7 +26,7 @@ public class TbStorageDetail implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 入库单编号
@@ -66,5 +66,5 @@ public class TbStorageDetail implements Serializable {
     /**
      * 入库数量
      */
-    private Integer storageQuantity;
+    private Long storageQuantity;
 }
