+++ src/main/java/com/eshop/entity/logistics/TbLogisticsDeclare.java
@@ -15,7 +15,7 @@ import lombok.ToString;
  * </p>
  *
  * @author logistics
- * @since 2025-02-11
+ * @since 2025-02-12
  */
 @Getter
 @Setter
@@ -26,12 +26,12 @@ public class TbLogisticsDeclare implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 物流方式_id
      */
-    private Integer modeId;
+    private Long modeId;
 
     /**
      * 申报品名(中)
