+++ src/main/java/com/eshop/entity/logistics/TbPackingRange.java
@@ -15,7 +15,7 @@ import lombok.ToString;
  * </p>
  *
  * @author logistics
- * @since 2025-02-11
+ * @since 2025-02-12
  */
 @Getter
 @Setter
@@ -26,12 +26,12 @@ public class TbPackingRange implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 包材id
      */
-    private Integer packingId;
+    private Long packingId;
 
     /**
      * 价格
