+++ src/main/java/com/eshop/entity/order/TbOrderInvoiceDetail.java
@@ -15,7 +15,7 @@ import lombok.ToString;
  * </p>
  *
  * @author order
- * @since 2025-02-11
+ * @since 2025-02-12
  */
 @Getter
 @Setter
@@ -26,7 +26,7 @@ public class TbOrderInvoiceDetail implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 发票编号
@@ -51,7 +51,7 @@ public class TbOrderInvoiceDetail implements Serializable {
     /**
      * 商品数量
      */
-    private Integer amount;
+    private Long amount;
 
     /**
      * 合计金额(数量*单价)
