+++ src/main/java/com/eshop/entity/order/TbOrderTypeMiddle.java
@@ -14,7 +14,7 @@ import lombok.ToString;
  * </p>
  *
  * @author order
- * @since 2025-02-11
+ * @since 2025-02-12
  */
 @Getter
 @Setter
@@ -25,15 +25,15 @@ public class TbOrderTypeMiddle implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 订单id
      */
-    private Integer orderId;
+    private Long orderId;
 
     /**
      * 订单类型id
      */
-    private Integer typeId;
+    private Long typeId;
 }
