+++ src/main/java/com/eshop/entity/order/TbWaybillSite.java
@@ -14,7 +14,7 @@ import lombok.ToString;
  * </p>
  *
  * @author order
- * @since 2025-02-11
+ * @since 2025-02-12
  */
 @Getter
 @Setter
@@ -25,12 +25,12 @@ public class TbWaybillSite implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 运单id
      */
-    private Integer waybillId;
+    private Long waybillId;
 
     /**
      * 买家姓
